Add .NET 9.0 support and EF Core 9.0 package reference
The project file `TnTEnumEntity.csproj` has been updated to include support for .NET 9.0. Specifically, the `TargetFrameworks` property now includes `net9.0` in addition to the existing `net6.0` and `net8.0`. Additionally, a new package reference for `Microsoft.EntityFrameworkCore` version `9.0.0-rc.1.24451.1` has been added, which is conditioned to be included only when the target framework is `net9.0`.